rename DBWrapper.ts -> SqliteDB.ts
