+++ packages/modeldb-sqlite/src/ModelDB.ts
@@ -12,7 +12,7 @@ import {
 	WhereCondition,
 } from "@canvas-js/modeldb"
 import { ModelAPI } from "@canvas-js/modeldb-sqlite-shared"
-import { SqliteDB } from "./DBWrapper.js"
+import { SqliteDB } from "./SqliteDB.js"
 
 export interface ModelDBOptions {
 	path: string | null
